Make fuzz tests run as part of unit tests
Moved content in fuzz.go into fuzz_test.go and removed references to
gofuzz build constraint.
Removed local_test.go as it doesn't do anything.
