+++ azure-pipelines.yml
@@ -56,7 +56,7 @@ jobs:
 
       - script: |
           set -e
-          go test -tags gofuzz -race -v -coverprofile=coverage.txt -covermode atomic ./... 2>&1 | go-junit-report > report.xml
+          go test -tags -race -v -coverprofile=coverage.txt -covermode atomic ./... 2>&1 | go-junit-report > report.xml
           gocov convert coverage.txt > coverage.json
           gocov-xml < coverage.json > coverage.xml
           gocov-html < coverage.json > coverage.html
